+++ src/MessageReceived.cpp
@@ -7,7 +7,7 @@
 
 MessageSent MessageReceived::empty_message_sent(0, Position(0,0,0,0,0,0,0,0,0));
 
-MessageReceived::MessageReceived(float simulation_t, int source_station_i, int origin_station_i):
+MessageReceived::MessageReceived(float simulation_t, long source_station_i, long origin_station_i):
 simulation_time(simulation_t), source_station_id(source_station_i) ,origin_station_id(origin_station_i), message_send(empty_message_sent){}
 
 void MessageReceived::set_message_sent(MessageSent &message_s) {
@@ -17,10 +17,10 @@ void MessageReceived::set_message_sent(MessageSent &message_s) {
 const float& MessageReceived::get_simulation_time() const{
     return simulation_time;
 }
-const int& MessageReceived::get_origin_station_id() const{
+const long& MessageReceived::get_origin_station_id() const{
     return origin_station_id;
 }
-const int& MessageReceived::get_source_station_id() const{
+const long& MessageReceived::get_source_station_id() const{
     return source_station_id;
 }
